put shortcut code in a single file
