+++ xpano/gui/panels/sidebar.cc
@@ -14,6 +14,7 @@
 #include "constants.h"
 #include "gui/action.h"
 #include "gui/panels/thumbnail_pane.h"
+#include "gui/shortcut.h"
 #include "utils/imgui_.h"
 
 namespace xpano::gui {
@@ -34,44 +35,16 @@ std::string ProgressLabel(algorithm::ProgressType type) {
   }
 }
 
-enum class ShortcutType { kOpen, kExport, kDebug };
-
-const char* Shortcut(ShortcutType type) {
-#if defined(__APPLE__)
-  switch (type) {
-    case ShortcutType::kOpen:
-      return reinterpret_cast<const char*>(u8"⌘ O");
-    case ShortcutType::kExport:
-      return reinterpret_cast<const char*>(u8"⌘ S");
-    case ShortcutType::kDebug:
-      return reinterpret_cast<const char*>(u8"⌘ D");
-    default:
-      return "";
-  }
-#else
-  switch (type) {
-    case ShortcutType::kOpen:
-      return "CTRL+O";
-    case ShortcutType::kExport:
-      return "CTRL+S";
-    case ShortcutType::kDebug:
-      return "CTRL+D";
-    default:
-      return "";
-  }
-#endif
-}
-
 Action DrawFileMenu() {
   Action action{};
   if (ImGui::BeginMenu("File")) {
-    if (ImGui::MenuItem("Open files", Shortcut(ShortcutType::kOpen))) {
+    if (ImGui::MenuItem("Open files", Label(ShortcutType::kOpen))) {
       action |= {ActionType::kOpenFiles};
     }
     if (ImGui::MenuItem("Open directory")) {
       action |= {ActionType::kOpenDirectory};
     }
-    if (ImGui::MenuItem("Export", Shortcut(ShortcutType::kExport))) {
+    if (ImGui::MenuItem("Export", Label(ShortcutType::kExport))) {
       action |= {ActionType::kExport};
     }
     ImGui::Separator();
@@ -127,7 +100,7 @@ Action DrawOptionsMenu(algorithm::CompressionOptions* compression_options,
 Action DrawHelpMenu() {
   Action action{};
   if (ImGui::BeginMenu("Help")) {
-    if (ImGui::MenuItem("Show debug info", Shortcut(ShortcutType::kDebug))) {
+    if (ImGui::MenuItem("Show debug info", Label(ShortcutType::kDebug))) {
       action |= {ActionType::kToggleDebugLog};
     }
     ImGui::Separator();
