+++ xpano/gui/shortcut.h
@@ -1,0 +1,14 @@
+
+#include <imgui.h>
+
+#include "gui/action.h"
+
+namespace xpano::gui {
+
+enum class ShortcutType { kOpen, kExport, kDebug };
+
+const char* Label(ShortcutType type);
+
+Action CheckKeybindings();
+
+}  // namespace xpano::gui
